requirements.txt
django>=3.2.14
oracledb
pillow
python-dotenv
python-magic-bin
